+++ packages/expo-video/android/src/main/java/expo/modules/video/player/VideoPlayer.kt
@@ -1,4 +1,4 @@
-package expo.modules.video
+package expo.modules.video.player
 
 import android.content.Context
 import android.view.SurfaceView
@@ -12,11 +12,13 @@ import androidx.media3.exoplayer.ExoPlayer
 import androidx.media3.ui.PlayerView
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.sharedobjects.SharedObject
+import expo.modules.video.VideoManager
 import expo.modules.video.delegates.IgnoreSameSet
 import expo.modules.video.enums.PlayerStatus
 import expo.modules.video.enums.PlayerStatus.*
 import expo.modules.video.playbackService.ExpoVideoPlaybackService
 import expo.modules.video.playbackService.PlaybackServiceConnection
+import expo.modules.video.records.BufferOptions
 import expo.modules.video.records.PlaybackError
 import expo.modules.video.records.VideoSource
 import expo.modules.video.records.VolumeEvent
@@ -29,11 +31,14 @@ class VideoPlayer(val context: Context, appContext: AppContext, source: VideoSou
   // This improves the performance of playing DRM-protected content
   private var renderersFactory = DefaultRenderersFactory(context)
     .forceEnableMediaCodecAsynchronousQueueing()
+    .setEnableDecoderFallback(true)
   private var listeners: MutableList<WeakReference<VideoPlayerListener>> = mutableListOf()
+  val loadControl: VideoPlayerLoadControl = VideoPlayerLoadControl.Builder().build()
 
   val player = ExoPlayer
     .Builder(context, renderersFactory)
     .setLooper(context.mainLooper)
+    .setLoadControl(loadControl)
     .build()
 
   val serviceConnection = PlaybackServiceConnection(WeakReference(player))
@@ -85,6 +90,12 @@ class VideoPlayer(val context: Context, appContext: AppContext, source: VideoSou
       sendEvent(PlayerEvent.PlaybackRateChanged(new.speed, old.speed))
     }
   }
+
+  var bufferOptions: BufferOptions = BufferOptions()
+    set(value) {
+      field = value
+      loadControl.applyBufferOptions(value)
+    }
 
   private val playerListener = object : Player.Listener {
     override fun onIsPlayingChanged(isPlaying: Boolean) {