+++ packages/expo-video/ios/VideoPlayer.swift
@@ -75,6 +75,17 @@ internal final class VideoPlayer: SharedRef<AVPlayer>, Hashable, VideoPlayerObse
     }
   }
 
+  var bufferOptions = BufferOptions() {
+    didSet {
+      pointer.currentItem?.preferredForwardBufferDuration = bufferOptions.preferredForwardBufferDuration
+      pointer.automaticallyWaitsToMinimizeStalling = bufferOptions.waitsToMinimizeStalling
+    }
+  }
+
+  var bufferedPosition: Double {
+    return getBufferedPosition()
+  }
+
   override init(_ pointer: AVPlayer) {
     observer = VideoPlayerObserver(player: pointer)
     super.init(pointer)
@@ -117,6 +128,7 @@ internal final class VideoPlayer: SharedRef<AVPlayer>, Hashable, VideoPlayerObse
     }
 
     playerItem.audioTimePitchAlgorithm = preservesPitch ? .spectral : .varispeed
+    playerItem.preferredForwardBufferDuration = bufferOptions.preferredForwardBufferDuration
     pointer.replaceCurrentItem(with: playerItem)
   }
 
@@ -135,6 +147,22 @@ internal final class VideoPlayer: SharedRef<AVPlayer>, Hashable, VideoPlayerObse
         track.isEnabled = enabled
       }
     })
+  }
+
+  private func getBufferedPosition() -> Double {
+    guard let currentItem = pointer.currentItem else {
+      return -1
+    }
+    let currentTime = pointer.currentTime().seconds
+
+    for timeRange in currentItem.loadedTimeRanges {
+      let start = CMTimeGetSeconds(timeRange.timeRangeValue.start)
+      let end = CMTimeGetSeconds(timeRange.timeRangeValue.end)
+      if start <= currentTime && end >= currentTime {
+        return end
+      }
+    }
+    return 0
   }
 
   // MARK: - VideoPlayerObserverDelegate
@@ -183,6 +211,7 @@ internal final class VideoPlayer: SharedRef<AVPlayer>, Hashable, VideoPlayerObse
 
   func onItemChanged(player: AVPlayer, oldVideoPlayerItem: VideoPlayerItem?, newVideoPlayerItem: VideoPlayerItem?) {
     safeEmit(event: "sourceChange", arguments: newVideoPlayerItem?.videoSource, oldVideoPlayerItem?.videoSource)
+    newVideoPlayerItem?.preferredForwardBufferDuration = bufferOptions.preferredForwardBufferDuration
   }
 
   func safeEmit<each A: AnyArgument>(event: String, arguments: repeat each A) {